Update registry structure
Renamed from did to registry
Introduced a types.go with interfaces and types
Moved all config related things to config.go
Removed obsolete engine tests
