+++ go.mod
@@ -8,8 +8,12 @@ require (
 	github.com/labstack/echo/v4 v4.1.17
 	github.com/lestrrat-go/jwx v1.0.7
 	github.com/magiconair/properties v1.8.4
-	github.com/nuts-foundation/go-did v0.0.0-20210114144203-c658fb4d0f65 // indirect
-	github.com/nuts-foundation/nuts-network v0.16.0 // indirect
+	github.com/nuts-foundation/go-did v0.0.0-20210114144203-c658fb4d0f65
+	github.com/nuts-foundation/nuts-crypto v0.16.0
+	github.com/nuts-foundation/nuts-go-core v0.16.0
+	github.com/nuts-foundation/nuts-go-test v0.16.0
+	github.com/nuts-foundation/nuts-network v0.16.0
+	github.com/nuts-foundation/nuts-registry v0.16.0
 	github.com/pkg/errors v0.9.1
 	github.com/prometheus/client_golang v1.7.1
 	github.com/sirupsen/logrus v1.7.0
