+++ registry/api/v1/generated.go
