+++ registry/logging/log.go
