+++ registry/logging/log_test.go
